feat(pipeline): refine progress tracking
