+++ server/common/thread_pool.py
@@ -14,6 +14,7 @@ def process_with_thread_pool(
     max_workers: int,
     timeout: int | float | None,
     on_error: Callable[[int, T, Exception], R],
+    on_progress: Callable[[int, int], None] | None = None,
 ) -> List[R]:
     """Process ``chunks`` in parallel with a thread pool.
 
@@ -35,6 +36,7 @@ def process_with_thread_pool(
     ex = ThreadPoolExecutor(max_workers=max_workers)
     futures = [ex.submit(func, i + 1, ch) for i, ch in enumerate(chunks)]
     try:
+        total = len(chunks)
         for i, fut in enumerate(futures, 1):
             try:
                 if timeout in (0, None):
@@ -46,6 +48,8 @@ def process_with_thread_pool(
             except Exception as e:  # pragma: no cover - passthrough to handler
                 res = on_error(i, chunks[i - 1], e)
             results.append(res)
+            if on_progress:
+                on_progress(i, total)
     except KeyboardInterrupt:  # pragma: no cover - user requested termination
         for f in futures:
             f.cancel()
